Relocate EscapeRoomCard component and ignore .env file
