--- .env
@@ -1,1 +1,0 @@
-VITE_API_URL=http://localhost:3000/
