+++ .gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 coverage
+.env
 
 # Editor directories and files
 .vscode/*
